initial
This is a small example how to provide a configuration to objects
defined in different Javascript files. RequireJS, makes it easy for us
to pass these objects to our files within the main page. This is how i
find great not to ever bother in assigning php values within the
javascript files.